backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
anthropic
openai
python-dotenv==1.0.1
secure==0.3.0